+++ src/brr.mli
@@ -1916,10 +1916,21 @@ module El : sig
   (** [find_by_tag_name ~root n] are the elements with tag name [t] that
       are descendents of [root] (defaults to {!Document.root}). *)
 
+  val find_first_by_selector : ?root:t -> Jstr.t -> t option
+  (** [find_first_by_selector ~root sel] is the
+      {{:https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelector}first element selected} by the CSS selector [sel] that is descendent
+      of [root] (defaults to {!Document.root}). *)
+
+  val fold_find_by_selector : ?root:t -> (t -> 'a -> 'a) -> Jstr.t -> 'a -> 'a
+  (** [fold_find_by_selector ~root f sel acc] folds [f] over the
+      {{:https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelectorAll}elements selected} by the CSS selector [sel] that are descendent of [root]
+      (defaults to {!Document.root}). *)
+
   (** {1:tree Parent and children} *)
 
   val parent : t -> t option
-  (** [parent e] is the {{:https://developer.mozilla.org/en-US/docs/Web/API/Node/parentNode}
+  (** [parent e] is the
+      {{:https://developer.mozilla.org/en-US/docs/Web/API/Node/parentNode}
       parent} {e element} of [e] (if any). *)
 
   val children : ?only_els:bool -> t -> t list